Shrink the database
In "Shrink the Database" Section when we attempt to launch SQLCMD using ./SQLCMD.EXE -S "(localdb)\.\ADSync" -U <Username> -P <Password>, the execution fails stating "Password: Sqlcmd: Error: Microsoft SQL Server Native Client 11.0 : Login failed
for user" and this is the case with the account used during installation or a user who is a member of the local administrators or Sync Service account or a user who is a member of local group ADSyncAdmins

If we do not specify -U and -P switch then execution completes successfully. So ideally SQLCMD should be executed as SQLCMD using ./SQLCMD.EXE -S "(localdb)\.\ADSync".

As per the article for SQLCMD https://docs.microsoft.com/en-us/sql/tools/sqlcmd-utility?view=sql-server-ver15. "If neither the -U option or the -P option is specified, sqlcmd tries to connect by using Microsoft Windows Authentication mode. Authentication is based on the Windows account of the user who is running sqlcmd"